+++ src/main/java/com/example/udpm14sellcomputerpartsbackend/model/dto/OrderDetailDto.java
@@ -8,7 +8,7 @@ import java.math.BigDecimal;
 public class OrderDetailDto {
     private Long id;
 
-    private BigDecimal price;
+    private long price;
 
     private Integer quantity;
